+++ app/src/main/java/ar/com/wolox/lucasdelatorre/training/activities/Login.java
@@ -39,14 +39,16 @@ public class Login extends Activity {
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
 
-        isLogged();
+        openBoard();
+
+        /*isLogged();
 
         setContentView(R.layout.activity_login);
 
         init();
         setUi();
         populate();
-        setListeners();
+        setListeners();*/
     }
 
     private void init() {
