+++ app/src/main/java/ar/com/wolox/lucasdelatorre/training/adapters/ViewPagerAdapter.java
@@ -1,11 +1,9 @@
-package ar.com.wolox.lucasdelatorre.training.supportabs;
+package ar.com.wolox.lucasdelatorre.training.adapters;
 
 import android.support.v4.app.Fragment;
-import android.support.v4.app.FragmentActivity;
 import android.support.v4.app.FragmentManager;
 import android.support.v4.app.FragmentStatePagerAdapter;
 
-import ar.com.wolox.lucasdelatorre.training.R;
 import ar.com.wolox.lucasdelatorre.training.fragments.News;
 import ar.com.wolox.lucasdelatorre.training.fragments.Profile;
