+++ app/src/main/java/ar/com/wolox/lucasdelatorre/training/instances/NewInstance.java
@@ -1,4 +1,4 @@
-package ar.com.wolox.lucasdelatorre.training;
+package ar.com.wolox.lucasdelatorre.training.instances;
 
 import java.util.ArrayList;
